third_party/blink/web_tests/svg/parser/whitespace-angle-invalid-2-expected.txt

CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+ ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ". ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "-.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".-grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "+.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, ".E0grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "e1grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " - ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " + ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " . ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " -.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .-grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " +.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " .E0grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, " e1grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   - ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   + ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   . ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   -.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .-grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   +.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   .E0grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "   e1grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t - ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t + ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t . ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t -.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .-grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t +.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t .E0grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\r\n\t e1grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1deg".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1rad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1grad".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+ ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.- ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+. ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1 ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1deg ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1rad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1grad ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1deg   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1rad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1grad   ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1deg\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1rad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1grad\r\n\t ".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f-.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.-grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f+.grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\f.E0grad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1deg\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1rad\f".
CONSOLE ERROR: Error: <marker> attribute orient: Expected angle, "\fe1grad\f".
This is a testharness.js-based test.
Harness: the test ran to completion.
